Imported google-checks into checkstyle & Removed unused classes
