--- src/main/java/pingis/services/DataImporterIO.java
@@ -1,21 +1,0 @@
-package pingis.services;
-
-import java.util.Scanner;
-
-public class DataImporterIO implements IO {
-
-    private Scanner s;
-    public DataImporterIO() {
-        s = new Scanner(getClass().getClassLoader().getResourceAsStream("exampledata/dummychallenges.json"));
-    }
-    @Override
-    public boolean hasNext() {
-        return s.hasNext();
-    }
-
-    @Override
-    public String nextLine() {
-        return s.nextLine();
-    }
-    
-}
